pyproject.toml
[tool.poetry]
name = "fas_contents_deploy"
version = "0.1.0"
description = ""
authors = ["Haruhiko Takada <haruhiko.takada@gmail.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.56"
markdown2 = "^2.4.6"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pyproject-flake8 = "^5.0.4"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
ignore = "E203"
exclude = ".venv"

[tool.isort]
profile = "black"